+++ mkdocs.yml
@@ -1,0 +1,1 @@
+site_name: My Docs
